backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
numpy
psycopg2
psycopg2-binary
django-cors-headers
redis
django-redis
pymongo
requests
web3
celery
channels
python-dotenv
transformers
Prophet
stable_baselines3
gym
#aave-python==0.2.0